# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Member system for dog-company (register, login, profile).",
        "title": "dog-company API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/leaderboard": {
            "get": {
                "description": "Returns top entries sorted by days ASC, then money DESC. Auth is optional — if a valid Bearer token is provided and the user has any entry for this goal, the response also includes my_best with the user's best entry and global rank. Default goal=50000, default limit=10 (max 50).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Get leaderboard top N",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "max entries (default 10, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "goal amount (default 50000)",
                        "name": "goal",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_leaderboard.ListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records a race-to-target completion. Nickname is taken from the authenticated user's account (client cannot spoof). Duplicate submissions (same user+goal+days+money within 1 minute) are silently deduplicated and return the existing entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Submit a leaderboard entry",
                "parameters": [
                    {
                        "description": "submit payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_leaderboard.SubmitPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_leaderboard.SubmitOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "bad payload / sanity failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "missing / invalid / expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    }
                }
            }
        },
        "/leaderboard/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's own entries sorted by submitted_at DESC. Optional goal filter. Default limit=20 (max 50).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Get my leaderboard history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "max entries (default 20, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "goal filter; omit for all goals",
                        "name": "goal",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_leaderboard.ListMineOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing / invalid / expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    }
                }
            }
        },
        "/saves": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's current save. If no save exists, data is null.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "saves"
                ],
                "summary": "Get current save",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_save.GetOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing / invalid / expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates the save if absent (accepts any client revision, server starts at 1). If present, client revision must equal server's current revision, otherwise 409 with server state. Body is capped at 64 KiB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "saves"
                ],
                "summary": "Create or overwrite current save",
                "parameters": [
                    {
                        "description": "save payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_save.SavePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_save.UpsertOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "bad payload / unsupported version / sanity failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "missing / invalid / expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "409": {
                        "description": "revision conflict; data carries server state",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_save.ConflictData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Idempotently deletes the authenticated user's save. Returns 200 even when no save existed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "saves"
                ],
                "summary": "Delete current save",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "missing / invalid / expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_leosu-cela_dog-company_pkg_tool.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_leosu-cela_dog-company_pkg_tool.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_leaderboard.EntryOutput": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "integer"
                },
                "goal": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "money": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "office_level": {
                    "type": "integer"
                },
                "staff_count": {
                    "type": "integer"
                },
                "submitted_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_leaderboard.ListMineOutput": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_leaderboard.EntryOutput"
                    }
                }
            }
        },
        "internal_leaderboard.ListOutput": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_leaderboard.EntryOutput"
                    }
                },
                "my_best": {
                    "$ref": "#/definitions/internal_leaderboard.MyBest"
                }
            }
        },
        "internal_leaderboard.MyBest": {
            "type": "object",
            "properties": {
                "entry": {
                    "$ref": "#/definitions/internal_leaderboard.EntryOutput"
                },
                "rank": {
                    "type": "integer"
                }
            }
        },
        "internal_leaderboard.SubmitOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_leaderboard.SubmitPayload": {
            "type": "object",
            "required": [
                "days",
                "goal",
                "money"
            ],
            "properties": {
                "days": {
                    "type": "integer",
                    "example": 58
                },
                "goal": {
                    "type": "integer",
                    "example": 50000
                },
                "money": {
                    "type": "integer",
                    "example": 52340
                },
                "office_level": {
                    "type": "integer",
                    "example": 4
                },
                "staff_count": {
                    "type": "integer",
                    "example": 9
                }
            }
        },
        "internal_save.ConflictData": {
            "type": "object",
            "properties": {
                "server_data": {
                    "type": "object"
                },
                "server_revision": {
                    "type": "integer"
                },
                "server_updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_save.GetOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "revision": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "internal_save.SavePayload": {
            "type": "object",
            "required": [
                "data",
                "version"
            ],
            "properties": {
                "data": {
                    "type": "object"
                },
                "revision": {
                    "type": "integer",
                    "example": 3
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_save.UpsertOutput": {
            "type": "object",
            "properties": {
                "revision": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\" (without quotes).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}